refactor(controller): extract service and status builders (part 7/7, final)
Part 7 of 7 of the god-controller split — the final PR.
Tracks #311.

Split the remaining single-concern chunks out of
inferenceservice_controller.go into two focused files:

  service_builder.go (~75 lines):
    (*InferenceServiceReconciler).constructService

  status_builder.go (~230 lines):
    (*InferenceServiceReconciler).reconcileVLLMSpecCondition
    (*InferenceServiceReconciler).constructEndpoint
    (*InferenceServiceReconciler).updateStatusWithSchedulingInfo

constructEndpoint lives with status_builder rather than service_builder
because its only caller is the status writer that populates
Status.Endpoint from the cluster-local service URL.

Also drops imports (fmt, meta, intstr) that are no longer referenced
in the controller file after this move.

Pure relocation — no signature changes, no behavior changes. Recreated
from current main rather than git-rebased due to overlapping conflict
regions across the accumulated prior splits.

End of the split. With all 7 parts landed, the controller file
contains only orchestration (Reconcile, reconcileDeployment/
reconcileService wrappers, getModelForInferenceService,
SetupWithManager, pod/model watches) plus small cross-cutting helpers
(sanitizeDNSName, boolPtr, initContainerSecurityContext).

Final line counts for the new layout:
  inferenceservice_controller.go: 356 (down from 1,567 — 77% smaller)
  deployment_builder.go:          229
  model_storage.go:               296
  hpa_reconciler.go:              195
  scheduling.go:                  201
  gpu_sharding.go:                126
  runtime_llamacpp_args.go:       130
  status_builder.go:              232
  service_builder.go:              75

Verified: go vet clean, golangci-lint 0 issues, make test green,
internal/controller coverage unchanged at 83.1%.
