+++ internal/controller/service_builder.go
@@ -1,0 +1,75 @@
+/*
+Copyright 2025.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package controller
+
+import (
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
+
+	inferencev1alpha1 "github.com/defilantech/llmkube/api/v1alpha1"
+)
+
+// constructService builds the Service fronting an InferenceService's pods.
+// Defaults to ClusterIP + port 8080; spec.endpoint.type ("NodePort" or
+// "LoadBalancer") upgrades the service type, and spec.endpoint.port overrides
+// the port. The endpoint URL that ends up on status is constructed separately
+// in status_builder.go.
+
+func (r *InferenceServiceReconciler) constructService(isvc *inferencev1alpha1.InferenceService) *corev1.Service {
+	serviceName := sanitizeDNSName(isvc.Name)
+
+	labels := map[string]string{
+		"app":                           isvc.Name,
+		"inference.llmkube.dev/service": isvc.Name,
+	}
+
+	port := int32(8080)
+	if isvc.Spec.Endpoint != nil && isvc.Spec.Endpoint.Port > 0 {
+		port = isvc.Spec.Endpoint.Port
+	}
+
+	serviceType := corev1.ServiceTypeClusterIP
+	if isvc.Spec.Endpoint != nil && isvc.Spec.Endpoint.Type != "" {
+		switch isvc.Spec.Endpoint.Type {
+		case "NodePort":
+			serviceType = corev1.ServiceTypeNodePort
+		case "LoadBalancer":
+			serviceType = corev1.ServiceTypeLoadBalancer
+		}
+	}
+
+	return &corev1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      serviceName,
+			Namespace: isvc.Namespace,
+			Labels:    labels,
+		},
+		Spec: corev1.ServiceSpec{
+			Type:     serviceType,
+			Selector: labels,
+			Ports: []corev1.ServicePort{
+				{
+					Name:       "http",
+					Port:       port,
+					TargetPort: intstr.FromInt(int(port)),
+					Protocol:   corev1.ProtocolTCP,
+				},
+			},
+		},
+	}
+}
